Removed separate controller for target allocation
