+++ pkg/collector/reconcile/deployment.go
@@ -25,6 +25,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	"github.com/open-telemetry/opentelemetry-operator/pkg/collector"
+	"github.com/open-telemetry/opentelemetry-operator/pkg/targetallocator"
 )
 
 // +kubebuilder:rbac:groups="apps",resources=deployments,verbs=get;list;watch;create;update;patch;delete
@@ -34,6 +35,14 @@ func Deployments(ctx context.Context, params Params) error {
 	desired := []appsv1.Deployment{}
 	if params.Instance.Spec.Mode == "deployment" {
 		desired = append(desired, collector.Deployment(params.Config, params.Log, params.Instance))
+	}
+
+	if params.Instance.Spec.TargetAllocator.Enabled {
+		_, err := GetPromConfig(params)
+		if err != nil {
+			return fmt.Errorf("failed to parse Prometheus config: %v", err)
+		}
+		desired = append(desired, targetallocator.Deployment(params.Config, params.Log, params.Instance))
 	}
 
 	// first, handle the create/update parts
@@ -79,7 +88,11 @@ func expectedDeployments(ctx context.Context, params Params, expected []appsv1.D
 			updated.Labels = map[string]string{}
 		}
 
-		updated.Spec = desired.Spec
+		if desired.Labels["app.kubernetes.io/component"] == "opentelemetry-targetallocator" {
+			updated.Spec.Template.Spec.Containers[0].Image = desired.Spec.Template.Spec.Containers[0].Image
+		} else {
+			updated.Spec = desired.Spec
+		}
 		updated.ObjectMeta.OwnerReferences = desired.ObjectMeta.OwnerReferences
 
 		for k, v := range desired.ObjectMeta.Annotations {
