+++ pkg/collector/reconcile/service.go
@@ -23,6 +23,7 @@ import (
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/intstr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
@@ -30,6 +31,7 @@ import (
 	"github.com/open-telemetry/opentelemetry-operator/pkg/collector"
 	"github.com/open-telemetry/opentelemetry-operator/pkg/collector/adapters"
 	"github.com/open-telemetry/opentelemetry-operator/pkg/naming"
+	"github.com/open-telemetry/opentelemetry-operator/pkg/targetallocator"
 )
 
 // +kubebuilder:rbac:groups="",resources=services,verbs=get;list;watch;create;update;patch;delete
@@ -46,6 +48,14 @@ func Services(ctx context.Context, params Params) error {
 				desired = append(desired, *svc)
 			}
 		}
+	}
+
+	if params.Instance.Spec.TargetAllocator.Enabled {
+		_, err := GetPromConfig(params)
+		if err != nil {
+			return fmt.Errorf("failed to parse Prometheus config: %v", err)
+		}
+		desired = append(desired, desiredTAService(params))
 	}
 
 	// first, handle the create/update parts
@@ -117,6 +127,30 @@ func desiredService(ctx context.Context, params Params) *corev1.Service {
 			Selector:  selector,
 			ClusterIP: "",
 			Ports:     ports,
+		},
+	}
+}
+
+func desiredTAService(params Params) corev1.Service {
+	labels := targetallocator.Labels(params.Instance)
+	labels["app.kubernetes.io/name"] = naming.TAService(params.Instance)
+
+	selector := targetallocator.Labels(params.Instance)
+	selector["app.kubernetes.io/name"] = naming.TargetAllocator(params.Instance)
+
+	return corev1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      naming.TAService(params.Instance),
+			Namespace: params.Instance.Namespace,
+			Labels:    labels,
+		},
+		Spec: corev1.ServiceSpec{
+			Selector: selector,
+			Ports: []corev1.ServicePort{{
+				Name:       "targetallocation",
+				Port:       443,
+				TargetPort: intstr.FromInt(443),
+			}},
 		},
 	}
 }
