+++ pkg/targetallocator/labels.go
@@ -31,7 +31,7 @@ func Labels(instance v1alpha1.OpenTelemetryCollector) map[string]string {
 	}
 
 	base["app.kubernetes.io/managed-by"] = "opentelemetry-operator"
-	base["app.kubernetes.io/instance"] = fmt.Sprintf("%s.%s", instance.Name, "targetallocator")
+	base["app.kubernetes.io/instance"] = fmt.Sprintf("%s.%s", instance.Namespace, instance.Name)
 	base["app.kubernetes.io/part-of"] = "opentelemetry"
 	base["app.kubernetes.io/component"] = "opentelemetry-targetallocator"
