+++ pkg/targetallocator/labels_test.go
@@ -35,7 +35,7 @@ func TestLabelsCommonSet(t *testing.T) {
 	// test
 	labels := Labels(otelcol)
 	assert.Equal(t, "opentelemetry-operator", labels["app.kubernetes.io/managed-by"])
-	assert.Equal(t, "my-instance.targetallocator", labels["app.kubernetes.io/instance"])
+	assert.Equal(t, "my-ns.my-instance", labels["app.kubernetes.io/instance"])
 	assert.Equal(t, "opentelemetry", labels["app.kubernetes.io/part-of"])
 	assert.Equal(t, "opentelemetry-targetallocator", labels["app.kubernetes.io/component"])
 }
