--- pkg/targetallocator/reconcile/suite_test.go
@@ -1,196 +1,0 @@
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package reconcile
-
-import (
-	"context"
-	"fmt"
-	"io/ioutil"
-	"os"
-	"path/filepath"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	"k8s.io/apimachinery/pkg/util/uuid"
-	"k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/envtest"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
-
-	"github.com/open-telemetry/opentelemetry-operator/api/v1alpha1"
-	"github.com/open-telemetry/opentelemetry-operator/internal/config"
-)
-
-var k8sClient client.Client
-var testEnv *envtest.Environment
-var testScheme *runtime.Scheme = scheme.Scheme
-var logger = logf.Log.WithName("unit-tests")
-
-var instanceUID = uuid.NewUUID()
-
-func TestMain(m *testing.M) {
-	testEnv = &envtest.Environment{
-		CRDDirectoryPaths: []string{filepath.Join("..", "..", "..", "config", "crd", "bases")},
-	}
-
-	cfg, err := testEnv.Start()
-	if err != nil {
-		fmt.Printf("failed to start testEnv: %v", err)
-		os.Exit(1)
-	}
-
-	if err := v1alpha1.AddToScheme(testScheme); err != nil {
-		fmt.Printf("failed to register scheme: %v", err)
-		os.Exit(1)
-	}
-	// +kubebuilder:scaffold:scheme
-
-	k8sClient, err = client.New(cfg, client.Options{Scheme: testScheme})
-	if err != nil {
-		fmt.Printf("failed to setup a Kubernetes client: %v", err)
-		os.Exit(1)
-	}
-
-	code := m.Run()
-
-	err = testEnv.Stop()
-	if err != nil {
-		fmt.Printf("failed to stop testEnv: %v", err)
-		os.Exit(1)
-	}
-
-	os.Exit(code)
-}
-
-func params() Params {
-	replicas := int32(1)
-	configYAML, err := ioutil.ReadFile("suite_test.yaml")
-	if err != nil {
-		fmt.Printf("Error getting yaml file: %v", err)
-	}
-	return Params{
-		Config: config.New(),
-		Client: k8sClient,
-		Instance: v1alpha1.OpenTelemetryCollector{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "opentelemetry.io",
-				APIVersion: "v1",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "test",
-				Namespace: "default",
-				UID:       instanceUID,
-			},
-			Spec: v1alpha1.OpenTelemetryCollectorSpec{
-				Mode: v1alpha1.ModeStatefulSet,
-				Ports: []v1.ServicePort{{
-					Name: "web",
-					Port: 80,
-					TargetPort: intstr.IntOrString{
-						Type:   intstr.Int,
-						IntVal: 80,
-					},
-					NodePort: 0,
-				}},
-				TargetAllocator: v1alpha1.OpenTelemetryTargetAllocator{
-					Enabled: true,
-				},
-				Replicas: &replicas,
-				Config:   string(configYAML),
-			},
-		},
-		Scheme: testScheme,
-		Log:    logger,
-	}
-}
-
-func newParams(containerImage ...string) Params {
-	replicas := int32(1)
-	configYAML, err := ioutil.ReadFile("suite_test.yaml")
-	if err != nil {
-		fmt.Printf("Error getting yaml file: %v", err)
-	}
-
-	cfg := config.New()
-	defaultContainerImage := cfg.TargetAllocatorImage()
-	if len(containerImage) > 0 && len(containerImage[0]) > 0 {
-		defaultContainerImage = containerImage[0]
-	}
-
-	return Params{
-		Config: cfg,
-		Client: k8sClient,
-		Instance: v1alpha1.OpenTelemetryCollector{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "opentelemetry.io",
-				APIVersion: "v1",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "test",
-				Namespace: "default",
-				UID:       instanceUID,
-			},
-			Spec: v1alpha1.OpenTelemetryCollectorSpec{
-				Mode: v1alpha1.ModeStatefulSet,
-				Ports: []v1.ServicePort{{
-					Name: "web",
-					Port: 80,
-					TargetPort: intstr.IntOrString{
-						Type:   intstr.Int,
-						IntVal: 80,
-					},
-					NodePort: 0,
-				}},
-				TargetAllocator: v1alpha1.OpenTelemetryTargetAllocator{
-					Enabled: true,
-					Image:   defaultContainerImage,
-				},
-				Replicas: &replicas,
-				Config:   string(configYAML),
-			},
-		},
-		Scheme: testScheme,
-		Log:    logger,
-	}
-}
-
-func createObjectIfNotExists(tb testing.TB, name string, object client.Object) {
-	tb.Helper()
-	err := k8sClient.Get(context.Background(), client.ObjectKey{Namespace: "default", Name: name}, object)
-	if errors.IsNotFound(err) {
-		err := k8sClient.Create(context.Background(),
-			object)
-		assert.NoError(tb, err)
-	}
-}
-
-func populateObjectIfExists(t testing.TB, object client.Object, namespacedName types.NamespacedName) (bool, error) {
-	t.Helper()
-	err := k8sClient.Get(context.Background(), namespacedName, object)
-	if errors.IsNotFound(err) {
-		return false, nil
-	}
-	if err != nil {
-		return false, err
-	}
-	return true, nil
-
-}
